+++ django_comments_xtd/conf/defaults.py
@@ -39,15 +39,15 @@ COMMENTS_XTD_SEND_HTML_EMAIL = True
 COMMENTS_XTD_THREADED_EMAILS = True
 
 
-# Define what app.models can have what commenting features.
+# Define what commenting features a pair app_label.model can have.
 # TODO: Put django-comments-xtd settings under a dictionary, and merge
 #       COMMENTS_XTD_MAX_THREAD_LEVEL_BY_APP_MODEL with this one.
 COMMENTS_XTD_APP_MODEL_PERMISSIONS = {
-    # 'articles.article': {
-    #     'allow_flagging': True,
-    #     'allow_feedback': True,
-    #     'show_feedback': True,
-    # }
+    'default': {
+        'allow_flagging': False,
+        'allow_feedback': False,
+        'show_feedback': False,
+    }
 }
 
 # Define a lambda function to return the user representation. Used by